backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
sqlalchemy==2.0.29
google-generativeai==0.3.2
redis==5.0.3
requests==2.31.0
